+++ plugins/fbinstant/src/Leaderboard.js
@@ -139,13 +139,15 @@ var Leaderboard = new Class({
      * 
      * The data is requested in an async call, so the result isn't available immediately.
      * 
-     * When the call completes this Leaderboard will emit the `setscore` event along with the score, any extra data and the name of the Leaderboard.
+     * When the call completes this Leaderboard will emit the `setscore` event along with the LeaderboardScore object and the name of the Leaderboard.
+     * 
+     * If the save fails the event will send `null` as the score value.
      *
      * @method Phaser.FacebookInstantGamesPlugin.Leaderboard#setScore
      * @since 3.13.0
      * 
      * @param {integer} score - The new score for the player. Must be a 64-bit integer number.
-     * @param {string} [data] - Metadata to associate with the stored score. Must be less than 2KB in size.
+     * @param {(string|any)} [data] - Metadata to associate with the stored score. Must be less than 2KB in size. If an object is given it will be passed to `JSON.stringify`.
      * 
      * @return {this} This Leaderboard instance.
      */
@@ -153,11 +155,27 @@ var Leaderboard = new Class({
     {
         if (data === undefined) { data = ''; }
 
+        if (typeof data === 'object')
+        {
+            data = JSON.stringify(data);
+        }
+
         var _this = this;
 
         this.ref.setScoreAsync(score, data).then(function (entry)
         {
-            _this.emit('setscore', entry.getScore(), entry.getExtraData(), _this.name);
+            if (entry)
+            {
+                var score = LeaderboardScore(entry);
+
+                _this.playerScore = score;
+    
+                _this.emit('setscore', score, _this.name);
+            }
+            else
+            {
+                _this.emit('setscore', null, _this.name);
+            }
 
         }).catch(function (e)
         {
@@ -173,6 +191,8 @@ var Leaderboard = new Class({
      * The data is requested in an async call, so the result isn't available immediately.
      * 
      * When the call completes this Leaderboard will emit the `getplayerscore` event along with the score and the name of the Leaderboard.
+     * 
+     * If the player has not yet saved a score, the event will send `null` as the score value, and `playerScore` will be set to `null` as well.
      *
      * @method Phaser.FacebookInstantGamesPlugin.Leaderboard#getPlayerScore
      * @since 3.13.0
@@ -185,11 +205,18 @@ var Leaderboard = new Class({
 
         this.ref.getPlayerEntryAsync().then(function (entry)
         {
-            var score = LeaderboardScore(entry);
+            if (entry)
+            {
+                var score = LeaderboardScore(entry);
 
-            _this.playerScore = score;
-
-            _this.emit('getplayerscore', score, _this.name);
+                _this.playerScore = score;
+    
+                _this.emit('getplayerscore', score, _this.name);
+            }
+            else
+            {
+                _this.emit('getplayerscore', null, _this.name);
+            }
 
         }).catch(function (e)
         {
@@ -204,7 +231,7 @@ var Leaderboard = new Class({
      * 
      * The data is requested in an async call, so the result isn't available immediately.
      * 
-     * When the call completes this Leaderboard will emit the `getplayerscore` event along with the score and the name of the Leaderboard.
+     * When the call completes this Leaderboard will emit the `getscores` event along with an array of LeaderboardScore entries and the name of the Leaderboard.
      *
      * @method Phaser.FacebookInstantGamesPlugin.Leaderboard#getScores
      * @since 3.13.0
@@ -231,6 +258,47 @@ var Leaderboard = new Class({
             });
 
             _this.emit('getscores', _this.scores, _this.name);
+
+        }).catch(function (e)
+        {
+            console.warn(e);
+        });
+
+        return this;
+    },
+
+    /**
+     * Retrieves a set of leaderboard entries, based on the current player's connected players (including the current player), ordered by local rank within the set of connected players.
+     * 
+     * The data is requested in an async call, so the result isn't available immediately.
+     * 
+     * When the call completes this Leaderboard will emit the `getconnectedscores` event along with an array of LeaderboardScore entries and the name of the Leaderboard.
+     *
+     * @method Phaser.FacebookInstantGamesPlugin.Leaderboard#getConnectedScores
+     * @since 3.16.0
+     * 
+     * @param {integer} [count=10] - The number of entries to attempt to fetch from the leaderboard. Currently, up to a maximum of 100 entries may be fetched per query.
+     * @param {integer} [offset=0] - The offset from the top of the leaderboard that entries will be fetched from.
+     * 
+     * @return {this} This Leaderboard instance.
+     */
+    getConnectedScores: function (count, offset)
+    {
+        if (count === undefined) { count = 10; }
+        if (offset === undefined) { offset = 0; }
+
+        var _this = this;
+
+        this.ref.getConnectedPlayerEntriesAsync().then(function (entries)
+        {
+            _this.scores = [];
+
+            entries.forEach(function (entry)
+            {
+                _this.scores.push(LeaderboardScore(entry));
+            });
+
+            _this.emit('getconnectedscores', _this.scores, _this.name);
 
         }).catch(function (e)
         {